src/test-requirements.txt
# This file is managed centrally.  If you find the need to modify this as a
# one-off, please don't.  Intead, consult #openstack-charms and ask about
# requirements management in charms via bot-control.  Thank you.
charm-tools>=2.4.4
coverage>=3.6
mock>=1.2
flake8>=2.2.4,<=2.4.1
stestr>=2.2.0
requests>=2.18.4
python-manilaclient
git+https://github.com/openstack-charmers/zaza.git#egg=zaza
git+https://github.com/openstack-charmers/zaza-openstack-tests.git#egg=zaza.openstack